Distribution independent prerequisites file
A shell script to automatically configure the build according to system (Linux or Darwin) and distribution (Fedora, Red Hat, CentOS, Ubuntu, SUSE, Debian), alleviating the need for separate files for each system. The script will install c++ make prerequisites for rpm and deb systems, and provide the user the option to install xerces-c and xsd via source, or dnf/yum/apt, or skip (for existing installations).

The script works well on Fedora, but needs further work to satisfy other distributions and Darwin.